media-info: mkv/mpegts: update
commit e8908f5aeef952566f6bccde743c7735d3f8c6ef
h264parse: Don't switch to passthrough on set_caps()

This commit changes the offset and DTS of the first frame
of the mkv h264 stream. The changes seem harmless and are not
relevant in the context of this file.

Also changes the number of frames pushed in the mpegts streams.