Yesterday.flac.media_info: Update for changed flacparse behavior
This sample flac file has headers in the order STREAMINFO(0),
SEEKTABLE(3), VORBISCOMMENT(4), PADDING(1). Previously, flacparse would
needlessly reorder this to STREAMINFO,VORBISCOMMENT,SEEKTABLE,PADDING.
I changed flacparse to avoid this unnecessary reordering because it was
breaking some other files, so the media_info needed to be updated to
reflect this.